remove playground & add ex
